+++ ssl/statem/statem_dtls.c
@@ -117,6 +117,7 @@
 #include <string.h>
 #include <stdio.h>
 #include "../ssl_locl.h"
+#include "statem_locl.h"
 #include <openssl/buffer.h>
 #include <openssl/rand.h>
 #include <openssl/objects.h>
